refactor: move enums to new file
