+++ packages/common/src/index.ts
@@ -1,3 +1,4 @@
+export * from './enums';
 export * from './types';
 export * from './utils';
 export * from './web3-config-provider';
