+++ packages/web3/src/hooks/useAccount.ts
@@ -1,9 +1,7 @@
-import type { UniversalWeb3ProviderInterface } from '@ant-design/web3-common';
-
 import useProvider from './useProvider';
 
 export default function useAccount() {
-  const { account } = useProvider() as UniversalWeb3ProviderInterface;
+  const { account } = useProvider();
 
   return {
     account,
