+++ packages/web3/src/hooks/useConnection.ts
@@ -1,9 +1,7 @@
-import type { UniversalWeb3ProviderInterface } from '@ant-design/web3-common';
-
 import useProvider from './useProvider';
 
 export default function useConnection() {
-  const { connect, disconnect } = useProvider() as UniversalWeb3ProviderInterface;
+  const { connect, disconnect } = useProvider();
 
   return {
     connect,
